requirements.txt
faiss-cpu
fastapi
uvicorn
sentence-transformers
python-dotenv
requests
